+++ systems/economy.js
@@ -1,3 +1,5 @@
+const { giveAchievement } = require("./achievementSystem")
+
 const rewards = {
  C:5,
  U:10,
@@ -14,6 +16,12 @@ function rewardKamas(user, rarity){
  const gain = rewards[rarity] || 0
 
  user.kamas += gain
+
+ if(user.kamas>=1000) giveAchievement(user,"kamas1000")
+ if(user.kamas>=5000) giveAchievement(user,"kamas5000")
+ if(user.kamas>=20000) giveAchievement(user,"kamas20000")
+ if(user.kamas>=100000) giveAchievement(user,"kamas100000")     
+ if(user.kamas>=1000000) giveAchievement(user,"kamas1000000")
 
  return gain
 }